Prepare mock netalloc plugin for unit testing.
