+++ plugins/netalloc/descriptor/addr_alloc.go
@@ -16,14 +16,13 @@ package descriptor
 
 import (
 	"fmt"
-	"net"
-	"strings"
 
 	"github.com/ligato/cn-infra/logging"
 
 	"github.com/ligato/vpp-agent/api/models/netalloc"
 	kvs "github.com/ligato/vpp-agent/plugins/kvscheduler/api"
 	"github.com/ligato/vpp-agent/plugins/netalloc/descriptor/adapter"
+	"github.com/ligato/vpp-agent/plugins/netalloc/utils"
 )
 
 const (
@@ -83,40 +82,11 @@ func (d *AddrAllocDescriptor) parseAddr(addrAlloc *netalloc.AddressAllocation) (
 	case netalloc.AddressType_IPV6_ADDR:
 		fallthrough
 	case netalloc.AddressType_IPV6_GW:
-		ipNet, err := ParseIPAddr(addrAlloc.Address)
+		ipNet, err := utils.ParseIPAddr(addrAlloc.Address)
 		if err != nil {
 			return nil, err
 		}
 		return &netalloc.AddrAllocMetadata{IPAddr: ipNet}, nil
 	}
 	return nil, fmt.Errorf("address of undefined type: %s", addrAlloc.Address)
-}
-
-// ParseIPAddr parses IP address from string.
-func ParseIPAddr(addr string) (ipNet *net.IPNet, err error) {
-	if strings.Contains(addr, "/") {
-		// IP with mask
-		ip, ipNet, err := net.ParseCIDR(addr)
-		if err != nil {
-			return nil, err
-		}
-		ipNet.IP = ip
-		return ipNet, nil
-	}
-
-	// IP without mask
-	defaultIpv4Mask := net.CIDRMask(32, 32)
-	defaultIpv6Mask := net.CIDRMask(128, 128)
-
-	ip := net.ParseIP(addr)
-	if ip == nil {
-		return nil, fmt.Errorf("invalid IP address: %s", addr)
-	}
-	if ip.To4() != nil {
-		// IPv4 address
-		return &net.IPNet{IP: ip.To4(), Mask: defaultIpv4Mask}, nil
-	}
-
-	// IPv6 address
-	return &net.IPNet{IP: ip, Mask: defaultIpv6Mask}, nil
 }
